% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_columns.R
\name{ensure_column_types}
\alias{ensure_column_types}
\title{Ensure that column types are not all characters}
\usage{
ensure_column_types(data)
}
\description{
Ensure that column types are not all characters. This sometimes
occurs if the machine re-initiates the column headers twice in the same
data set.
}
\keyword{internal}
