% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise_prepost}
\alias{normalise_prepost}
\title{Normalise Breathing Traces}
\usage{
normalise_prepost(data, data_pre, data_post, method = "means_linear")
}
\arguments{
\item{data_post}{}
}
\description{
Normalise Breathing Traces
}
