% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_licor.R
\name{read_li7000}
\alias{read_li7000}
\title{Read from LI-7000}
\usage{
read_li7000(filepath)
}
\arguments{
\item{filepath}{Path to file}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
