% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respirometr-package.R
\docType{package}
\name{respirometr-package}
\alias{respirometr}
\alias{respirometr-package}
\title{respirometr: An R toolbox for analysing respirometry data}
\description{
An R toolbox for analysing respirometry data.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.roald-arboel.com/respirometr/}
  \item \url{https://github.com/roaldarbol/respirometr/}
  \item Report bugs at \url{https://github.com/roaldarbol/respirometr/issues}
}

}
\author{
\strong{Maintainer}: Mikkel Roald-Arbøl \email{respirometr.c2ymk@passmail.net} (\href{https://orcid.org/0000-0002-9998-0058}{ORCID})

}
\keyword{internal}
