% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise_prepost}
\alias{normalise_prepost}
\title{Normalise Traces}
\usage{
normalise_prepost(data, pre, post, method = "means_linear")
}
\arguments{
\item{data}{Data from trial}

\item{pre}{Pre-trial level. Case be either a single value (if known) or a vector of values, from which the mean is calculated.}

\item{post}{Post-trial level. Case be either a single value (if known) or a vector of values, from which the mean is calculated.}

\item{method}{Which method is used to normalise the data}
}
\value{
Normalised data
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
