% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise_prepost}
\alias{normalise_prepost}
\title{Normalise Breathing Traces}
\usage{
normalise_prepost(data, data_pre, data_post, method = "means_linear")
}
\arguments{
\item{data}{Data from trial}

\item{data_pre}{Data, pre-trial}

\item{data_post}{Data, post-trial}

\item{method}{Which method is used to normalise the data}
}
\value{
Normalised data
}
\description{
Normalise Breathing Traces
}
