% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_licor.R
\name{read_licor}
\alias{read_licor}
\title{Read Licor data}
\usage{
read_licor(filepath, model = "LI7000")
}
\arguments{
\item{filepath}{Path to file}

\item{model}{LiCor model}
}
\value{
Tidy LiCor data
}
\description{
Read Licor data
}
